# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: '0.0.0.0:8080'
basePath: /api
tags:
  - name: method-controller
    description: Method Controller
  - name: method-level-controller
    description: Method Level Controller
  - name: method-type-controller
    description: Method Type Controller
paths:
  /methods:
    get:
      tags:
        - method-controller
      summary: getAllMethods
      operationId: getAllMethodsUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: X-User-ID
          in: header
          description: X-User-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Method'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - method-controller
      summary: createMethod
      operationId: createMethodUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: X-User-ID
          in: header
          description: X-User-ID
          required: true
          type: string
        - in: body
          name: methodResource
          description: methodResource
          required: true
          schema:
            $ref: '#/definitions/MethodResource'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MethodResource'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /methods/levels:
    get:
      tags:
        - method-level-controller
      summary: getAllMethodLevels
      operationId: getAllMethodLevelsUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MethodLevelResource'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - method-level-controller
      summary: createMethodLevel
      operationId: createMethodLevelUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: methodLevelResource
          description: methodLevelResource
          required: true
          schema:
            $ref: '#/definitions/MethodLevelResource'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MethodLevelResource'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/methods/levels/{methodLevelId}':
    put:
      tags:
        - method-level-controller
      summary: updateMethodLevel
      operationId: updateMethodLevelUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: methodLevelId
          in: path
          description: methodLevelId
          required: true
          type: string
        - in: body
          name: methodLevelResource
          description: methodLevelResource
          required: true
          schema:
            $ref: '#/definitions/MethodLevelResource'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MethodLevelResource'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /methods/types:
    get:
      tags:
        - method-type-controller
      summary: getAllMethodTypes
      operationId: getAllMethodTypesUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MethodTypeResource'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - method-type-controller
      summary: createMethodType
      operationId: createMethodTypeUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: methodTypeResource
          description: methodTypeResource
          required: true
          schema:
            $ref: '#/definitions/MethodTypeResource'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MethodTypeResource'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/methods/types/{methodTypeId}':
    put:
      tags:
        - method-type-controller
      summary: updateMethodType
      operationId: updateMethodTypeUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: methodTypeId
          in: path
          description: methodTypeId
          required: true
          type: string
        - in: body
          name: methodTypeResource
          description: methodTypeResource
          required: true
          schema:
            $ref: '#/definitions/MethodTypeResource'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MethodTypeResource'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/methods/{methodId}':
    get:
      tags:
        - method-controller
      summary: getMethod
      operationId: getMethodUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: methodId
          in: path
          description: methodId
          required: true
          type: string
        - name: X-User-ID
          in: header
          description: X-User-ID
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MethodResource'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - method-controller
      summary: updateMethod
      operationId: updateMethodUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: methodId
          in: path
          description: methodId
          required: true
          type: string
        - name: X-User-ID
          in: header
          description: X-User-ID
          required: true
          type: string
        - in: body
          name: methodResource
          description: methodResource
          required: true
          schema:
            $ref: '#/definitions/MethodResource'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MethodResource'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - method-controller
      summary: deleteMethod
      operationId: deleteMethodUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: methodId
          in: path
          description: methodId
          required: true
          type: string
        - name: X-User-ID
          in: header
          description: X-User-ID
          required: true
          type: string
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
definitions:
  Attachment:
    type: object
    properties:
      content:
        type: string
      created_at:
        type: string
        example: 'yyyy-MM-dd''T''HH:mm:ss.SSSz'
      created_by:
        $ref: '#/definitions/User'
      id:
        type: string
      modified_at:
        type: string
        example: 'yyyy-MM-dd''T''HH:mm:ss.SSSz'
      modified_by:
        $ref: '#/definitions/User'
      title:
        type: string
  Method:
    type: object
    properties:
      attachments:
        type: array
        items:
          $ref: '#/definitions/Attachment'
      content:
        type: string
      created_at:
        type: string
        format: date-time
      created_by:
        $ref: '#/definitions/User'
      id:
        type: string
      method_levels:
        type: array
        items:
          $ref: '#/definitions/MethodLevel'
      method_types:
        type: array
        items:
          $ref: '#/definitions/MethodType'
      modified_at:
        type: string
        format: date-time
      modified_by:
        $ref: '#/definitions/User'
      seminar_goals:
        type: array
        items:
          type: string
      seminar_type:
        type: string
      title:
        type: string
  MethodLevel:
    type: object
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
  MethodLevelResource:
    type: object
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
  MethodResource:
    type: object
    properties:
      attachments:
        type: array
        items:
          $ref: '#/definitions/Attachment'
      content:
        type: string
      created_at:
        type: string
        format: date-time
      created_by:
        $ref: '#/definitions/User'
      id:
        type: string
      method_levels:
        type: array
        items:
          $ref: '#/definitions/MethodLevel'
      method_types:
        type: array
        items:
          $ref: '#/definitions/MethodType'
      modified_at:
        type: string
        format: date-time
      modified_by:
        $ref: '#/definitions/User'
      seminar_goals:
        type: array
        items:
          type: string
      seminar_type:
        type: string
      title:
        type: string
  MethodType:
    type: object
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
  MethodTypeResource:
    type: object
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
  User:
    type: object
    properties:
      id:
        type: string
      role:
        type: string
        enum:
          - ADMIN
          - EDITOR
          - WRITER
          - READER